Changed percentile's value type to BigDecimal

+++ app-core/src/main/kotlin/tech/testra/reportal/model/Models.kt
@@ -154,5 +154,5 @@ data class SimulationScenario(
 
 data class Percentile(
     val n: Byte,
-    val value: Long
+    val value: BigDecimal
 )
